Add files via upload
Crochet bookmarks are handmade, decorative pieces used to mark pages in a book, crafted using various crochet techniques. These bookmarks come in many designs, from simple to intricate, and are typically thin and lightweight to easily slip between the pages without damaging the book. Crochet bookmarks are cherished for their beauty, uniqueness, and the personal touch they add to reading.Crochet bookmarks are designed to be slim and flat so they can be placed between the pages of a book without adding bulk or bending the pages. This makes them practical for frequent readers.